requirements.txt
python-telegram-bot==20.7
httpx>=0.26
python-dotenv>=1.0
